# internal/graphql/model/models_gen.go
// Code generated by github.com/99designs/gqlgen, DO NOT EDIT.

package model

type Category struct {
	ID          int    `json:"id"`
	HouseholdID int    `json:"householdID"`
	Name        string `json:"name"`
	Icon        string `json:"icon"`
	CreatedAt   string `json:"createdAt"`
	UpdatedAt   string `json:"updatedAt"`
}

type CategorySummary struct {
	CategoryID   int    `json:"categoryID"`
	CategoryName string `json:"categoryName"`
	Recurring    string `json:"recurring"`
	OneTime      string `json:"oneTime"`
	Total        string `json:"total"`
}

type CreateCategoryInput struct {
	HouseholdID int     `json:"householdID"`
	Name        string  `json:"name"`
	Icon        *string `json:"icon,omitempty"`
}

type CreateHouseholdInput struct {
	Name        string  `json:"name"`
	Description *string `json:"description,omitempty"`
	Currency    string  `json:"currency"`
	Icon        *string `json:"icon,omitempty"`
}

type CreateRecurringExpenseInput struct {
	HouseholdID int     `json:"householdID"`
	CategoryID  int     `json:"categoryID"`
	Name        string  `json:"name"`
	Description *string `json:"description,omitempty"`
	Amount      string  `json:"amount"`
	Frequency   string  `json:"frequency"`
	StartDate   string  `json:"startDate"`
	EndDate     *string `json:"endDate,omitempty"`
}

type CreateScheduleOverrideInput struct {
	RecurringExpenseID int    `json:"recurringExpenseID"`
	EffectiveDate      string `json:"effectiveDate"`
	Amount             string `json:"amount"`
	Frequency          string `json:"frequency"`
}

type CreateTransactionInput struct {
	HouseholdID int     `json:"householdID"`
	CategoryID  int     `json:"categoryID"`
	Amount      string  `json:"amount"`
	Description *string `json:"description,omitempty"`
	Date        string  `json:"date"`
}

type Household struct {
	ID          int    `json:"id"`
	Name        string `json:"name"`
	Description string `json:"description"`
	Currency    string `json:"currency"`
	Icon        string `json:"icon"`
	OwnerID     int    `json:"ownerID"`
	CreatedAt   string `json:"createdAt"`
	UpdatedAt   string `json:"updatedAt"`
}

type MonthlySummary struct {
	Month             string            `json:"month"`
	HouseholdID       int               `json:"householdID"`
	TotalIncome       string            `json:"totalIncome"`
	TotalExpenses     string            `json:"totalExpenses"`
	RecurringTotal    string            `json:"recurringTotal"`
	RecurringIncome   string            `json:"recurringIncome"`
	RecurringExpenses string            `json:"recurringExpenses"`
	OneTimeTotal      string            `json:"oneTimeTotal"`
	OneTimeIncome     string            `json:"oneTimeIncome"`
	OneTimeExpenses   string            `json:"oneTimeExpenses"`
	MonthlyTotal      string            `json:"monthlyTotal"`
	CategoryBreakdown []CategorySummary `json:"categoryBreakdown"`
}

type Mutation struct {
}

type Query struct {
}

type RecurringExpense struct {
	ID          int     `json:"id"`
	HouseholdID int     `json:"householdID"`
	CategoryID  int     `json:"categoryID"`
	Name        string  `json:"name"`
	Description string  `json:"description"`
	Amount      string  `json:"amount"`
	Frequency   string  `json:"frequency"`
	Active      bool    `json:"active"`
	StartDate   string  `json:"startDate"`
	EndDate     *string `json:"endDate,omitempty"`
	CreatedAt   string  `json:"createdAt"`
	UpdatedAt   string  `json:"updatedAt"`
}

type ScheduleOverride struct {
	ID                 int    `json:"id"`
	RecurringExpenseID int    `json:"recurringExpenseID"`
	EffectiveDate      string `json:"effectiveDate"`
	Amount             string `json:"amount"`
	Frequency          string `json:"frequency"`
	CreatedAt          string `json:"createdAt"`
	UpdatedAt          string `json:"updatedAt"`
}

type Transaction struct {
	ID          int    `json:"id"`
	HouseholdID int    `json:"householdID"`
	CategoryID  int    `json:"categoryID"`
	Amount      string `json:"amount"`
	Description string `json:"description"`
	Date        string `json:"date"`
	CreatedAt   string `json:"createdAt"`
	UpdatedAt   string `json:"updatedAt"`
}

type UpdateCategoryInput struct {
	ID   int     `json:"id"`
	Name string  `json:"name"`
	Icon *string `json:"icon,omitempty"`
}

type UpdateHouseholdInput struct {
	ID          int     `json:"id"`
	Name        string  `json:"name"`
	Description *string `json:"description,omitempty"`
	Currency    string  `json:"currency"`
	Icon        *string `json:"icon,omitempty"`
}

type UpdateRecurringExpenseInput struct {
	ID          int     `json:"id"`
	CategoryID  int     `json:"categoryID"`
	Name        string  `json:"name"`
	Description *string `json:"description,omitempty"`
	Amount      string  `json:"amount"`
	Frequency   string  `json:"frequency"`
	Active      bool    `json:"active"`
	StartDate   string  `json:"startDate"`
	EndDate     *string `json:"endDate,omitempty"`
}

type UpdateScheduleOverrideInput struct {
	ID            int    `json:"id"`
	EffectiveDate string `json:"effectiveDate"`
	Amount        string `json:"amount"`
	Frequency     string `json:"frequency"`
}

type UpdateTransactionInput struct {
	ID          int     `json:"id"`
	HouseholdID int     `json:"householdID"`
	CategoryID  int     `json:"categoryID"`
	Amount      string  `json:"amount"`
	Description *string `json:"description,omitempty"`
	Date        string  `json:"date"`
}