# web/static/openapi.yaml
openapi: 3.0.1
info:
  title: Money Tracker API
  description: REST API for the Money Tracker household budget application.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token obtained via POST /api/v1/tokens
  schemas:
...
    Summary:
      type: object
      properties:
        month:
          type: string
          example: "2026-01"
        household_id:
          type: integer
        total_income:
          type: string
          example: "3000"
        total_expenses:
          type: string
          example: "-75"
        recurring_total:
          type: string
          example: "-800"
        recurring_income:
          type: string
          example: "2000"
        recurring_expenses:
          type: string
          example: "-800"
        one_time_total:
          type: string
          example: "-75"
        one_time_income:
          type: string
          example: "1000"
        one_time_expenses:
          type: string
          example: "-75"
        monthly_total:
          type: string
          example: "-875"
        category_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
    CategorySummary:
      type: object
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        recurring:
          type: string
        one_time:
          type: string
        total:
          type: string
...
  parameters:
    householdId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    month:
      name: month
      in: query
      required: true
      schema:
        type: string
        pattern: '^\d{4}-\d{2}$'
        example: "2026-01"
paths:
...
  /households/{id}/summary:
    get:
      summary: Get monthly summary
      operationId: getMonthlySummary
      tags: [Summary]
      parameters:
        - $ref: '#/components/parameters/householdId'
        - $ref: '#/components/parameters/month'
      responses:
        '200':
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '400':
          description: Invalid household ID or month
        '401':
          description: Unauthorized
        '404':
          description: Household not found
...